remove _data and investigate
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

+++ torchao/float8/float8_tensor.py
@@ -166,7 +166,7 @@ class _ToFloat8ConstrFunc(torch.autograd.Function):
         # Required by scaled_mm, scale is always float32.
         # Cast tensor to float32 to improve numerics and
         # get on-par with torch.compile.
-        tensor_scaled = tensor.to(torch.float32) * scale
+        tensor_scaled = tensor * scale
         bits_fp8 = to_fp8_saturated(tensor_scaled, float8_dtype)
 
         if isinstance(bits_fp8, DTensor):
